Convert BatteryStats.kt to a protobuf
