+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/database/WatchBatteryStats.kt
@@ -10,15 +10,15 @@ import java.util.UUID
 data class WatchBatteryStats(
     @PrimaryKey val watchId: UUID,
     @ColumnInfo(name = "watch_battery_percent")
-    override val percent: Int,
+    val percent: Int,
     @ColumnInfo(name = "watch_charging")
-    override val isCharging: Boolean,
+    val charging: Boolean,
     @ColumnInfo(name = "last_update_time")
-    override val lastUpdatedMillis: Long = System.currentTimeMillis()
-) : BatteryStats(percent, isCharging, lastUpdatedMillis) {
+    val lastUpdatedMillis: Long = System.currentTimeMillis()
+) {
     companion object {
         fun BatteryStats.toWatchBatteryStats(watchId: UUID): WatchBatteryStats {
-            return WatchBatteryStats(watchId, percent, isCharging, lastUpdatedMillis)
+            return WatchBatteryStats(watchId, percent, charging, timestamp)
         }
     }
 }
